requirements.txt
python-telegram-bot==21.9
motor==3.3.2
pymongo==4.6.1
chromadb>=0.5.0
pydantic>=2.5
pydantic-settings>=2.1
google-generativeai>=0.8
anthropic>=0.40
openai>=1.0
pypdf>=4.0
python-docx>=1.1
beautifulsoup4>=4.12
httpx>=0.27
